+++ examples/admin-sdk-app/src/frontend/cms/ex-dailymotion/ex-dailymotion-element.vue
@@ -20,7 +20,7 @@
 
 <script setup lang="ts">
 import { ref, computed, onBeforeMount } from 'vue';
-import { data } from '@shopware-ag/meteor-admin-sdk';
+import { data } from '@xchanming-ag/meteor-admin-sdk';
 import EX_DAILYMOTION_CONSTANTS from './ex-dailymotion-constants';
 
 const dataSelectors = [