+++ examples/admin-sdk-app/src/frontend/locations/ex-product-extension-example-resize.vue
@@ -29,7 +29,7 @@
 
 <script setup lang="ts">
 import { ref, computed } from "vue";
-import { location } from "@shopware-ag/meteor-admin-sdk";
+import { location } from "@xchanming-ag/meteor-admin-sdk";
 import { SwButton, SwNumberField } from '@shopware-ag/meteor-component-library';
 
 const heightInput = ref(123);