+++ examples/admin-sdk-plugin/src/Resources/app/administration/src/base/mainCommands.ts
@@ -1,4 +1,4 @@
-import { notification, ui, context } from '@shopware-ag/meteor-admin-sdk';
+import { notification, ui, context } from '@xchanming-ag/meteor-admin-sdk';
 
 // location/index
 ui.componentSection.add({