+++ examples/admin-sdk-plugin/src/Resources/app/administration/src/views/location/index.ts
@@ -1,5 +1,5 @@
 import { defineComponent } from 'vue'
-import { location } from "@shopware-ag/meteor-admin-sdk";
+import { location } from "@xchanming-ag/meteor-admin-sdk";
 import { MtButton, MtNumberField } from '@shopware-ag/meteor-component-library';
 
 export default defineComponent({