+++ examples/admin-sdk-plugin/src/Resources/app/administration/src/views/ui/modals/modal-content.ts
@@ -1,5 +1,5 @@
 import { defineComponent } from 'vue'
-import { ui } from '@shopware-ag/meteor-admin-sdk';
+import { ui } from '@xchanming-ag/meteor-admin-sdk';
 
 export default defineComponent({
     template: `