+++ examples/admin-sdk-plugin/src/Resources/app/administration/src/views/ui/modals/modals.ts
@@ -1,5 +1,5 @@
 import { defineComponent } from 'vue'
-import { ui } from '@shopware-ag/meteor-admin-sdk';
+import { ui } from '@xchanming-ag/meteor-admin-sdk';
 
 // make background color white
 document.body.style.backgroundColor = 'white';