+++ examples/admin-sdk-plugin/src/Resources/app/administration/CHANGELOG.md
@@ -32,7 +32,7 @@
 - Updated dependencies [3863b64]
 - Updated dependencies [fab1a1d]
   - @shopware-ag/meteor-component-library@4.1.0
-  - @shopware-ag/meteor-admin-sdk@5.7.0
+  - @xchanming-ag/meteor-admin-sdk@5.7.0
 
 ## 1.0.2
 
@@ -40,7 +40,7 @@
 
 - Updated dependencies [f3b0e2f]
   - @shopware-ag/meteor-component-library@4.0.1
-  - @shopware-ag/meteor-admin-sdk@5.6.2
+  - @xchanming-ag/meteor-admin-sdk@5.6.2
 
 ## 1.0.1
 
@@ -78,4 +78,4 @@
 - Updated dependencies [50de30f]
 - Updated dependencies [b3039c1]
   - @shopware-ag/meteor-component-library@4.0.0
-  - @shopware-ag/meteor-admin-sdk@5.6.1
+  - @xchanming-ag/meteor-admin-sdk@5.6.1